add : force user list update after follow or unfollow
